requirements.txt
Django>=1.8,<2.0
djangorestframework
psycopg2
